Adds a heartbeat to prevent connection timeouts (#4)
The timeout feature is great, but it is easy to trigger a timeout if no output has occurred in a while. So instead of waiting for output all the time, this PR adds a HEARTBEAT command, which is sent by the server while a task is pending.

It makes no attempt to ensure that the task is making progress, but it does ensure that the connection is kept alive.