Delay queued notification message
The client aggressively prints “Command RUN is queued” when it recieves a heartbeat quickly but the command is still in the queued state. This happens regardless of whether there is anything else in the queue, so its kind of a noisy red-herring. This PR delays the appearance of that messge.